requirements.txt
pyTelegramBotAPI
google-generativeai
schedule
pillow
flask